requirements.txt
numpy>=1.17.4
pillow>=6.2.0
matplotlib>=3.1.1
click>=7.0
